Move Analytics, Config and deep link Push commands into :core module
