+++ core/build.gradle
@@ -16,6 +16,7 @@ android {
 dependencies {
   implementation project(':core-api')
   implementation project(':core-android-api')
+  implementation project(':feature:push-api')
 
   implementation 'androidx.appcompat:appcompat:1.5.1'
   implementation 'androidx.core:core:1.9.0'
