+++ core/src/main/java/com/jraska/github/client/core/push/ConfigAsPropertyCommand.kt
@@ -1,8 +1,11 @@
-package com.jraska.github.client.push
+package com.jraska.github.client.core.push
 
 import com.jraska.github.client.Config
 import com.jraska.github.client.Owner
 import com.jraska.github.client.analytics.AnalyticsProperty
+import com.jraska.github.client.push.PushAction
+import com.jraska.github.client.push.PushActionCommand
+import com.jraska.github.client.push.PushExecuteResult
 import javax.inject.Inject
 
 internal class ConfigAsPropertyCommand @Inject constructor(
